reordering function args
