+++ libcudacxx/include/cuda/__annotated_ptr/createpolicy.h
@@ -27,71 +27,73 @@
 
 _LIBCUDACXX_BEGIN_NAMESPACE_CUDA
 
-enum class _L2_Policy
+enum class __l2_evict_t : uint32_t
 {
-  __evict_last,
-  __evict_normal,
-  __evict_first,
-  __evict_unchanged
+  _L2_Evict_Unchanged     = 0, // called "_L2_Evict_Normal" at lower level
+  _L2_Evict_First         = 1,
+  _L2_Evict_Last          = 2,
+  _L2_Evict_Normal_Demote = 3
 };
 
 /***********************************************************************************************************************
  * PTX MAPPING
  **********************************************************************************************************************/
 
+#if _CCCL_HAS_CUDA_COMPILER()
+
 template <typename = void>
 [[nodiscard]] _CCCL_CONST _CCCL_HIDE_FROM_ABI _CCCL_DEVICE uint64_t __createpolicy_range_ptx(
-  size_t __gmem_ptr, _L2_Policy __primary, _L2_Policy __secondary, uint32_t __primary_size, uint32_t __total_size)
+  __l2_evict_t __primary, __l2_evict_t __secondary, size_t __gmem_ptr, uint32_t __primary_size, uint32_t __total_size)
 {
   uint64_t __policy;
-  if (__secondary == _L2_Policy::__evict_unchanged)
+  if (__secondary == __l2_evict_t::_L2_Evict_Unchanged)
   {
-    if (__primary == _L2_Policy::__evict_last)
+    if (__primary == __l2_evict_t::_L2_Evict_Last)
     {
       asm("createpolicy.range.global.L2::evict_last.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
-    else if (__primary == _L2_Policy::__evict_normal)
+    else if (__primary == __l2_evict_t::_L2_Evict_Normal_Demote)
     {
       asm("createpolicy.range.global.L2::evict_normal.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
-    else if (__primary == _L2_Policy::__evict_first)
+    else if (__primary == __l2_evict_t::_L2_Evict_First)
     {
       asm("createpolicy.range.global.L2::evict_first.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
-    else if (__primary == _L2_Policy::__evict_unchanged)
+    else if (__primary == __l2_evict_t::_L2_Evict_Unchanged)
     {
       asm("createpolicy.range.global.L2::evict_unchanged.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
   }
-  else // __secondary == __evict_first
+  else // __secondary == _L2_Evict_First
   {
-    if (__primary == _L2_Policy::__evict_last)
+    if (__primary == __l2_evict_t::_L2_Evict_Last)
     {
       asm("createpolicy.range.global.L2::evict_last.L2::evict_first.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
-    else if (__primary == _L2_Policy::__evict_normal)
+    else if (__primary == __l2_evict_t::_L2_Evict_Normal_Demote)
     {
       asm("createpolicy.range.global.L2::evict_normal.L2::evict_first.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
-    else if (__primary == _L2_Policy::__evict_first)
+    else if (__primary == __l2_evict_t::_L2_Evict_First)
     {
       asm("createpolicy.range.global.L2::evict_first.L2::evict_first.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
           : "l"(__gmem_ptr), "r"(__primary_size), "r"(__total_size));
     }
-    else if (__primary == _L2_Policy::__evict_unchanged)
+    else if (__primary == __l2_evict_t::_L2_Evict_Unchanged)
     {
       asm("createpolicy.range.global.L2::evict_unchanged.L2::evict_first.b64 %0, [%1], %2, %3;"
           : "=l"(__policy)
@@ -103,49 +105,51 @@ template <typename = void>
 
 template <typename = void>
 [[nodiscard]] _CCCL_CONST _CCCL_HIDE_FROM_ABI _CCCL_DEVICE uint64_t
-__createpolicy_fraction_ptx(_L2_Policy __primary, _L2_Policy __secondary, float __fraction)
+__createpolicy_fraction_ptx(__l2_evict_t __primary, __l2_evict_t __secondary, float __fraction)
 {
   uint64_t __policy;
-  if (__secondary == _L2_Policy::__evict_unchanged)
+  if (__secondary == __l2_evict_t::_L2_Evict_Unchanged)
   {
-    if (__primary == _L2_Policy::__evict_last)
+    if (__primary == __l2_evict_t::_L2_Evict_Last)
     {
       asm("createpolicy.fractional.L2::evict_last.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
-    else if (__primary == _L2_Policy::__evict_normal)
+    else if (__primary == __l2_evict_t::_L2_Evict_Normal_Demote)
     {
       asm("createpolicy.fractional.L2::evict_normal.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
-    else if (__primary == _L2_Policy::__evict_first)
+    else if (__primary == __l2_evict_t::_L2_Evict_First)
     {
       asm("createpolicy.fractional.L2::evict_first.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
-    else if (__primary == _L2_Policy::__evict_unchanged)
+    else if (__primary == __l2_evict_t::_L2_Evict_Unchanged)
     {
       asm("createpolicy.fractional.L2::evict_unchanged.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
   }
-  else // __secondary == __evict_first
+  else // __secondary == _L2_Evict_First
   {
-    if (__primary == _L2_Policy::__evict_last)
+    if (__primary == __l2_evict_t::_L2_Evict_Last)
     {
       asm("createpolicy.fractional.L2::evict_last.L2::evict_first.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
-    else if (__primary == _L2_Policy::__evict_normal)
+    else if (__primary == __l2_evict_t::_L2_Evict_Normal_Demote)
     {
       asm("createpolicy.fractional.L2::evict_normal.L2::evict_first.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
-    else if (__primary == _L2_Policy::__evict_first)
+    else if (__primary == __l2_evict_t::_L2_Evict_First)
     {
       asm("createpolicy.fractional.L2::evict_first.L2::evict_first.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
-    else if (__primary == _L2_Policy::__evict_unchanged)
+    else if (__primary == __l2_evict_t::_L2_Evict_Unchanged)
     {
       asm("createpolicy.fractional.L2::evict_unchanged.L2::evict_first.b64 %0, %1;" : "=l"(__policy) : "f"(__fraction));
     }
   }
   return __policy;
 }
+
+#endif // _CCCL_HAS_CUDA_COMPILER()
 
 /***********************************************************************************************************************
  * C++ API
@@ -155,29 +159,25 @@ extern "C" _CCCL_DEVICE void __createpolicy_is_not_supported_before_SM_80();
 
 template <typename T = void>
 [[nodiscard]] _CCCL_CONST _CCCL_HIDE_FROM_ABI _CCCL_DEVICE uint64_t __createpolicy_range(
-  const void* __ptr,
-  _L2_Policy __primary,
-  uint32_t __primary_size,
-  uint32_t __total_size,
-  _L2_Policy __secondary = _L2_Policy::__evict_unchanged)
+  __l2_evict_t __primary, __l2_evict_t __secondary, const void* __ptr, uint32_t __primary_size, uint32_t __total_size)
 {
   _CCCL_ASSERT(__isGlobal(__ptr), "ptr must be global");
   _CCCL_ASSERT(__primary_size <= __total_size, "primary_size must be less than or equal to total_size");
-  _CCCL_ASSERT(__secondary == _L2_Policy::__evict_first || __secondary == _L2_Policy::__evict_unchanged,
+  _CCCL_ASSERT(__secondary == __l2_evict_t::_L2_Evict_First || __secondary == __l2_evict_t::_L2_Evict_Unchanged,
                "secondary policy must be evict_first or evict_unchanged");
   auto __gmem_ptr = __cvta_generic_to_global(__ptr);
   NV_IF_ELSE_TARGET(
     NV_PROVIDES_SM_80,
-    (return ::cuda::__createpolicy_range_ptx(__gmem_ptr, __primary, __secondary, __primary_size, __total_size);),
+    (return ::cuda::__createpolicy_range_ptx(__primary, __secondary, __gmem_ptr, __primary_size, __total_size);),
     (__createpolicy_is_not_supported_before_SM_80(); return 0;))
 }
 
 template <typename T = void>
-[[nodiscard]] _CCCL_CONST _CCCL_HIDE_FROM_ABI _CCCL_DEVICE uint64_t __createpolicy_fraction(
-  _L2_Policy __primary, float __fraction = 1.0f, _L2_Policy __secondary = _L2_Policy::__evict_unchanged)
+[[nodiscard]] _CCCL_CONST _CCCL_HIDE_FROM_ABI _CCCL_DEVICE uint64_t
+__createpolicy_fraction(__l2_evict_t __primary, __l2_evict_t __secondary, float __fraction = 1.0f)
 {
   _CCCL_ASSERT(__fraction > 0.0f && __fraction <= 1.0f, "fraction must be between 0.0f and 1.0f");
-  _CCCL_ASSERT(__secondary == _L2_Policy::__evict_first || __secondary == _L2_Policy::__evict_unchanged,
+  _CCCL_ASSERT(__secondary == __l2_evict_t::_L2_Evict_First || __secondary == __l2_evict_t::_L2_Evict_Unchanged,
                "secondary policy must be evict_first or evict_unchanged");
   NV_IF_ELSE_TARGET(NV_PROVIDES_SM_80,
                     (return ::cuda::__createpolicy_fraction_ptx(__primary, __secondary, __fraction);),
